refactor: inquiry 컴포넌트 위치 재조정 (#105)
