+++ src/app/support/inquiry/_components/NoInquiryResult.tsx
@@ -1,7 +1,7 @@
 export const NoInquiryResult = () => {
   return (
-    <div className='border-gray flex h-[265px] w-[1400px] flex-col items-center justify-center rounded-lg border'>
-      <p className='text-xl font-medium'>
+    <div className='border-gray flex h-[265px] w-[1400px] flex-col items-center justify-center rounded-lg'>
+      <p className='text-center text-xl font-medium'>
         아직 남기신 문의가 없어요.
         <br />
         필요한 사항이 있다면 언제든 편하게 문의해주세요!
